requirements.txt
bleak==0.22.3
colorzero==2.0
dbus-fast==2.24.3
evdev==1.7.1
gpiozero==2.0.1
keyboard==0.13.5
libusb1==3.1.0
mfrc522==0.0.7
ndeflib==0.3.3
nfcpy==1.0.4
paho-mqtt==2.1.0
pi-rc522==2.3.0
pigpio==1.78
pyDes==2.0.1
pynput==1.7.7
pyserial==3.5
python-dotenv==1.0.1
python-xlib==0.33
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.5.0
spidev==3.6
typing_extensions==4.12.2
